Tidy up new Reconciliation Consistency check
