+++ BudgetAnalyser.Engine/Ledger/IReconciliationConsistency.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace BudgetAnalyser.Engine.Ledger
+{
+    public interface IReconciliationConsistency
+    {
+        IDisposable EnsureConsistency(LedgerBook book);
+    }
+}
